segal/client/requirements.txt
#asn1crypto==0.24.0
#certifi==2018.1.18
#cffi==1.15.1
#chardet==3.0.4
#cryptography==2.1.4
#idna==2.6
lxml==4.9.1
ntplib==0.3.4
#pycparser==2.18
pyOpenSSL==22.0.0
requests==2.18.4
#six==1.11.0
